feat: add scroll-into-view functionality
BREAKING CHANGE:
This adds a scroll-into-view functionality to the Link component.
Before these changes, the overflow behavior of the Nav component was
not set. Now, by default, the overflow behavior is set to 'auto,' and
the active Link component scrolls into the relative viewport of the Nav
component.

Other Changes:

bug-fix: fix the multi touch crash bug

This fixes the bug that causes the app to crash when multiple touches
occur on the Content component.

refractor: clean the console logs

This deletes all debug logs.